[SYSTEMML-370] Moved functions.terms tests to functions.misc
Closes #55.
